app/requirements.txt
fastapi
uvicorn
openai
youtube-transcript-api
pydub
edge-tts
python-dotenv
python-multipart